Refactor components from app component
