+++ youtube-downloader-webui/src/app/app.component.scss
@@ -2,45 +2,4 @@
   padding: 1rem;
   display: flex;
   flex-direction: column;
-
-  .form {
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    grid-template-rows: fit-content() repeat(4, 1fr);
-    column-gap: 1rem;
-
-    .search {
-      grid-column: 1 / span 2;
-    }
-
-    .title {
-      grid-row: 2;
-      grid-column: 1 / span 2;
-    }
-
-    .download-btn {
-      justify-self: start;
-    }
-  }
-
-  .download-container {
-    margin-top: 1rem;
-    display: grid;
-    grid-template-columns: 1fr 1fr;
-    gap: 1rem;
-  }
-}
-
-.progress {
-  display: flex;
-  margin-top: 0.5rem;
-  &-number {
-    margin-left: auto;
-  }
-  &-text.success {
-    color: green;
-  }
-  &-text.error {
-    color: red;
-  }
 }
